+++ src/redux/index.js
@@ -1,0 +1,5 @@
+import * as loginActions from './login/actions';
+
+export {
+  loginActions,
+};
