+++ src/scenes/home/Home.view.js
@@ -1,10 +1,8 @@
 import React from 'react';
+import { ActivityIndicator, View, Text, TouchableOpacity } from 'react-native';
 import PropTypes from 'prop-types';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import { View, Text, TouchableOpacity } from 'react-native';
-import * as loginActions from '../../redux/login/actions';
 import styles from './Home.styles';
+import { Colors } from '../../utils';
 
 class Home extends React.Component {
   login = () => {
@@ -13,16 +11,24 @@ class Home extends React.Component {
       password: 'reactnative',
     });
   };
+
   render() {
     return (
       <View style={styles.container}>
         <View>
           <Text style={styles.lead}>A simple react native boilerplate with a login action.</Text>
+          <ActivityIndicator animating={this.props.isLoading} size="large" color={Colors.primary} />
           {this.props.token && (
-            <View style={styles.token}>
+            <View style={styles.success}>
               <Text style={styles.tokenText}>You are now logged in. Your token: {this.props.token}</Text>
             </View>
           )}
+          {this.props.serverError && (
+            <View style={styles.error}>
+              <Text style={styles.tokenText}>There went something wrong: {this.props.serverError}</Text>
+            </View>
+          )}
+
         </View>
         <TouchableOpacity style={styles.button} onPress={this.login}>
           <Text style={styles.buttonText}>Fire a login action!</Text>
@@ -45,15 +51,5 @@ Home.defaultProps = {
   token: null,
 };
 
-const mapStateToProps = state => ({
-  isLoading: state.login.isLoading,
-  isLoggedIn: state.login.isLoggedIn,
-  serverError: state.login.serverError,
-  token: state.login.token,
-});
 
-const mapDispatchToProps = dispatch => ({
-  login: bindActionCreators(loginActions.login, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(Home);
+export default Home;
