+++ src/scenes/index.js
@@ -1,4 +1,4 @@
-import Home from './home/Home';
+import Home from './home/Home.container';
 
 export {
   Home,
